requirements.txt
Django~=4.1.1
djangorestframework~=3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==3.13.0
django-admin-interface==0.24.2
django-admin-sortable2~=2.1.3
django-summernote==0.8.20.0
django-pdfkit==0.3.1
django-faicon==0.1.0
django-rest-passwordreset==1.3.0
django-clone==5.3.0